FAI-552 - Update UpdatedPermissionEventHandler
Updated the handler to no longer call the accounts database to get the public encoded account legal entity id. Instead just gets it from encoding the known internal id. This is to stop the issue of the dashboard not being created when the ALE has been deleted and provider permissions updated