FAI-1931 - Change so get all apprenticeships comes from outer api
This caches the response from the outer api until the next day, instead of caching the query store response. It should be noted at the time of doing this the ICache implementation is a memory cache implementation only. This will be ok for this implementation as each deployed node can have its own version