FAI-2772 Some clean-up and fixing the bug in the PRE_PROD, issue related to the check your answers page validation
Fix:
They should be able to add new dates in the future AND proceed to the Check your answers page, where the validation will be displayed.
For this validation they then go to the select the course page and the provider can either select a listed course OR go to the Managing standards section and add it to their list.